Rename: change controller.js file route
