+++ src/js/index.js
@@ -1,7 +1,7 @@
 import '../scss/main.scss';
 import { createHeaderElement } from './views/headerView.js';
 import { createMainElement } from './views/mainView.js';
-import { controller } from './controller';
+import { controller } from './controller/controller.js';
 
 (() => {
   const headerView = createHeaderElement();
